Custom data source -> MatTableDataSource for better UI updates
