+++ Frontend/src/app/components/vacation-request-table/vacation-request-table.component.html
@@ -1,5 +1,5 @@
 <div class="mat-elevation-z8">
-  <table mat-table class="full-width-table" matSort aria-label="Elements">
+  <table mat-table class="full-width-table" matSort aria-label="Elements" [dataSource]="dataSource">
     <!-- Id Column -->
     <ng-container matColumnDef="nr">
       <th mat-header-cell *matHeaderCellDef>Nr</th>
@@ -41,10 +41,10 @@
   </table>
 
   <mat-paginator #paginator
-      [length]="dataSource.data.length"
-      [pageIndex]="0"
-      [pageSize]="10"
-      [pageSizeOptions]="[5, 10, 20]"
-      aria-label="Select page">
+                 [length]="dataSource.data.length"
+                 [pageIndex]="0"
+                 [pageSize]="10"
+                 [pageSizeOptions]="[5, 10, 20]"
+                 aria-label="Select page">
   </mat-paginator>
 </div>
